code division : extract file tab control from code.c
